+++ src/Scene/Objects/Camera.cpp
@@ -53,7 +53,7 @@ void Camera::lookAt(double x, double y, double z) {
     lookAt(double3(x, y, z));
 }
 
-Ray Camera::rayThroughPixel(const size3& window_size, uint16_t x, uint16_t y) const {
+Ray Camera::rayThroughPixel(const size3& window_size, size_t x, size_t y) const {
     double s_x = (x + 0.5) / window_size.x;
     double s_y = (y + 0.5) / window_size.y;
